+++ NEWS.md
@@ -1,6 +1,7 @@
 # BGData 2.1.0
 
 - Add `chunkedMap()` function.
+- Improve error handling in `chunkedMap()` and `chunkedApply()`.
 - `GWAS()`: Fix bug when computing p-values for methods other than rayOLS,
   lsfit, or SKAT when `i` is used to subset samples.